timestamp values (in microseconds) sent in feedback packets by armassist/rehand applications are now converted in python code to units of seconds, so that s-->us and us-->s conversions dont need to occur all over the place; next commit will delete all old commented code related to this change

+++ riglib/ismore/__init__.py
@@ -45,13 +45,13 @@ class ArmAssistData(FeedbackData):
 
     state_names = ['aa_px', 'aa_py', 'aa_ppsi']
     sub_dtype_data     = np.dtype([(name, np.float64) for name in state_names])
-    sub_dtype_ts       = np.dtype([(name, np.int64)   for name in state_names])
+    sub_dtype_ts       = np.dtype([(name, np.float64) for name in state_names])
     sub_dtype_data_aux = np.dtype([(name, np.float64) for name in ['force', 'bar_angle']])
-    sub_dtype_ts_aux   = np.dtype([(name, np.int64)   for name in ['force', 'bar_angle']])
+    sub_dtype_ts_aux   = np.dtype([(name, np.float64) for name in ['force', 'bar_angle']])
     
     dtype = np.dtype([('data',       sub_dtype_data),
                       ('ts',         sub_dtype_ts),
-                      ('ts_arrival', np.int64),
+                      ('ts_arrival', np.float64),
                       ('freq',       np.float64),
                       ('data_aux',   sub_dtype_data_aux),
                       ('ts_aux',     sub_dtype_ts_aux)])
@@ -78,15 +78,15 @@ class ReHandData(FeedbackData):
     state_names = ['rh_pthumb', 'rh_pindex', 'rh_pfing3', 'rh_pprono', 
                    'rh_vthumb', 'rh_vindex', 'rh_vfing3', 'rh_vprono']
     sub_dtype_data   = np.dtype([(name, np.float64) for name in state_names])
-    sub_dtype_ts     = np.dtype([(name, np.int64)   for name in state_names])
+    sub_dtype_ts     = np.dtype([(name, np.float64) for name in state_names])
     sub_dtype_torque = np.dtype([(name, np.float64) for name in ['thumb', 'index', 'fing3', 'prono']])
     
     dtype = np.dtype([('data',       sub_dtype_data),
                       ('ts',         sub_dtype_ts),
-                      ('ts_arrival', np.int64),
+                      ('ts_arrival', np.float64),
                       ('freq',       np.float64),
                       ('torque',     sub_dtype_torque),
-                      ('ts_sent',    np.int64)])
+                      ('ts_sent',    np.float64)])
 
     def get(self):
         d = self.data.next()
